# *: Move ShouldDelegateList into storage layer and make it exportable