@wojtek-t 2022-07-20
Updated release notes - it's not something our users are interested in.