@MadhavJivrajani 2022-07-26
@wojtek-t I've added a conversion function. PTAL

Also I'm not entirely sure what you meant here:
> Conversion func would help, although I would like it to be constructed in a way that it could also be used in:
https://github.com/kubernetes/kubernetes/blob/master/staging/src/k8s.io/apiserver/pkg/registry/generic/registry/store.go#L356

Could you please elaborate in case I misunderstood?

---

As per the benchmark, I wrote and ran the following benchmark:
```go
func BenchmarkShouldDelegateList(b *testing.B) {
	listOptions := metav1.ListOptions{
		ResourceVersion:      "12345",
		ResourceVersionMatch: metav1.ResourceVersionMatchNotOlderThan,
		Limit:                45,
		Continue:             strings.Repeat("a", 64),
	}
	b.ResetTimer()
	b.ReportAllocs()

	for i := 0; i < b.N; i++ {
                // Or run with the previous vanilla function that was deleted.
		storage.ShouldDelegateList(storage.ConvertToListOptions(listOptions))
	}
}
```

The benchmark was run for a period of 5s, 10 times over and then the results passed through `benchstat`:

```
❯ benchstat vanilla.txt conversion.txt
name                   old time/op    new time/op    delta
ShouldDelegateList-16    15.9ns ± 1%    28.1ns ± 1%  +76.35%  (p=0.000 n=9+9)

name                   old alloc/op   new alloc/op   delta
ShouldDelegateList-16     0.00B          0.00B          ~     (all equal)

name                   old allocs/op  new allocs/op  delta
ShouldDelegateList-16      0.00           0.00          ~     (all equal)
```
There is some cpu cost added (as expected), although I don't think this is anything significant considering the ns scale its in.
Lmk what you think, thanks!