@wojtek-t 2022-07-26
> Could you please elaborate in case I misunderstood?

Sorry. What I meant is that I would like this conversion function to also be used in the place I pointed at.
In such case, any problem with that function should surface pretty quickly, because it's *the* place that does that.

Regarding benchmark - can you please check that in to this PR?